x64编译通过
git-svn-id: http://qlanguage.googlecode.com/svn/trunk@167 6d982bb3-a5a7-393d-ffc5-582415be9c85

+++ Source/QCore/Library/graph/bitmap.cpp
@@ -62,8 +62,8 @@ NAMESPACE_QLANGUAGE_LIBRARY_START
             FILE* fp = fopen(path.c_str(), "wb");
             if (fp == NULL) throw error<const char*>("error open file", __FILE__, __LINE__);
             _BITMAPFILEHEADER fileHeader;
-            _BITMAPINFOHEADER infoHeader(width, height, type, sizePreLine * height);
-            fileHeader.size = sizeof(_BITMAPFILEHEADER) + sizeof(_BITMAPINFOHEADER) + sizePreLine * height;
+            _BITMAPINFOHEADER infoHeader((long)width, (long)height, type, (ulong)(sizePreLine * height));
+            fileHeader.size = (ulong)(sizeof(_BITMAPFILEHEADER) + sizeof(_BITMAPINFOHEADER) + sizePreLine * height);
             fileHeader.offset = sizeof(_BITMAPFILEHEADER) + sizeof(_BITMAPINFOHEADER);
             size_t quadCount = 0;
             switch (type)
